Separate tests between unit tests and integration tests
