+++ src/Polly.Caching.Distributed.SharedSpecs/Polly.Caching.Distributed.SharedSpecs.projitems
@@ -9,10 +9,10 @@
     <Import_RootNamespace>Polly.Caching.Distributed.Specs</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)NetStandardIDistributedCacheProvider_CacheRoundTripSpecs.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CacheProviderHelperTests.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)NetStandardIDistributedCacheByteArrayProviderSpecs.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)NetStandardIDistributedCacheStringProviderSpecs.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)TtlStrategyHelperTests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Integration\NetStandardIDistributedCacheProvider_CacheRoundTripSpecs.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Unit\CacheProviderHelperTests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Unit\NetStandardIDistributedCacheByteArrayProviderSpecs.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Unit\NetStandardIDistributedCacheStringProviderSpecs.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Unit\TtlStrategyHelperTests.cs" />
   </ItemGroup>
 </Project>
