+++ src/Polly.Caching.Distributed.SharedSpecs/Unit/CacheProviderHelperTests.cs
@@ -4,7 +4,7 @@ using Moq;
 using Polly.Caching.Distributed;
 using Xunit;
 
-namespace Polly.Specs.Caching.Distributed
+namespace Polly.Specs.Caching.Distributed.Unit
 {
     public class CacheProviderHelperTests
     {
